requirements.txt
flask==2.2.2
sqlalchemy==1.4.43
pydantic==1.10.2
uvicorn==0.19.0
gunicorn==20.1.0
pyyaml==6.0
typer==0.7.0
pytz==2022.6
flask_session==0.4.0
flask_apscheduler==1.12.4
ipython==8.6.0